+++ site/source/get_api_items.py
@@ -87,7 +87,7 @@ def exportItems():
 def main():
     parser = optparse.OptionParser(usage="Usage: %prog [options] version")
     parser.add_option("-s", "--siteapi", dest="siteapi", default="http://www.developer.nokia.com/Community/Wiki/api.php", help="Location of API")
-    (options, args) = parser.parse_args()
+    _options, _args = parser.parse_args()
     # print 'Site: %s' % options.siteapi
     parseFiles()
     exportItems()